webnn/matmul14.html

<!doctype html>
<meta charset=utf-8>
<title>test matmul operation</title>
<link rel="help" href="https://webmachinelearning.github.io/webnn/#api-mlgraphbuilder-matmul">
<script src=../resources/testharness.js></script>
<script src=../resources/testharnessreport.js></script>
<script src="./webnn-polyfill.js"></script>
<div id=log></div>
<script type="module">
  'use strict';
  import * as utils from './utils.js';
  let builder;
  setup(() => {
    const context = navigator.ml.createContext();
    builder = new MLGraphBuilder(context);
  })

  function testMatmul(A, B, expected) {
    const a = builder.input('a', {type: 'float32', dimensions: A.shape});
    const b = builder.constant(
        {type: 'float32', dimensions: B.shape}, new Float32Array(B.value));
    const c = builder.matmul(a, b);
    const graph = builder.build({c});
    const inputs = {'a': new Float32Array(A.value)};
    const outputs = {'c': new Float32Array(utils.sizeOfShape(expected.shape))};
    graph.compute(inputs, outputs);
    assert_array_approx_equals_ulp('matmul', outputs.c, expected.value, 8);
  }

  test(() => {
    testMatmul(
        {shape: [4], value: [87.03187561035156, 46.8813362121582, 95.82432556152344, -52.54386901855469]},
        {shape: [2, 2, 4, 2], value: [87.01571655273438, -86.33499145507812, 97.28276824951172, -94.84310150146484, -48.726898193359375, -20.1142520904541, 62.6322021484375, -47.690345764160156, -83.0711441040039, -59.670536041259766, -92.00804138183594, -44.09381103515625, 76.05716705322266, -24.82939338684082, 47.4744873046875, 48.12596893310547, -95.73054504394531, 24.911312103271484, -94.13774871826172, 55.28803634643555, 34.24235534667969, 60.54612731933594, -39.49250793457031, -27.449230194091797, 39.054264068603516, 71.23308563232422, 41.59154510498047, -25.262161254882812, 4.529367446899414, -7.762320518493652, -61.278778076171875, 88.5433120727539]},
        {shape: [2, 2, 1, 2], value: [
          4173.7265625,
      -11381.8671875,
      -6749.6640625,
      -12168.400390625,
      -7388.57275390625,
      12004.1357421875,
      9002.6806640625,
      -381.00213623046875]});
  }, 'matmul 1d x 4d');


</script>
